# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MyPets",
        "description": "bark",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "OpenApi host"
        }
    ],
    "paths": {
...
        "/auth/sign-in": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "description": "Login method",
                "operationId": "login",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "User email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "User password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Access token, pet owner and pet info and if not done registration - current step of registration.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "registration_process": {
                                            "type": "boolean"
                                        },
                                        "current_step": {
                                            "type": "integer"
                                        },
                                        "user": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "owner": {
                                                    "properties": {
                                                        "first_name": {
                                                            "type": "string"
                                                        },
                                                        "last_name": {
                                                            "type": "string"
                                                        },
                                                        "gender": {
                                                            "type": "string"
                                                        },
                                                        "age": {
                                                            "type": "integer"
                                                        },
                                                        "birthday": {
                                                            "type": "date"
                                                        },
                                                        "occupation": {
                                                            "type": "string"
                                                        },
                                                        "hobbies": {
                                                            "type": "string"
                                                        },
                                                        "pets_owned": {
                                                            "type": "string"
                                                        },
                                                        "profile_picture": {
                                                            "type": "string"
                                                        },
                                                        "favorite_park": {
                                                            "type": "string"
                                                        },
                                                        "pet": {
                                                            "properties": {
                                                                "id": {
                                                                    "type": "integer"
                                                                },
                                                                "name": {
                                                                    "type": "string"
                                                                },
                                                                "gender": {
                                                                    "type": "string"
                                                                },
                                                                "size": {
                                                                    "type": "string"
                                                                },
                                                                "primary_breed": {
                                                                    "type": "string"
                                                                },
                                                                "secondary_breed": {
                                                                    "type": "string"
                                                                },
                                                                "age": {
                                                                    "type": "integer"
                                                                },
                                                                "profile_picture": {
                                                                    "type": "string"
                                                                },
                                                                "friendliness": {
                                                                    "type": "integer"
                                                                },
                                                                "activity_level": {
                                                                    "type": "integer"
                                                                },
                                                                "noise_level": {
                                                                    "type": "integer"
                                                                },
                                                                "odebience_level": {
                                                                    "type": "integer"
                                                                },
                                                                "fetchability": {
                                                                    "type": "integer"
                                                                },
                                                                "swimability": {
                                                                    "type": "integer"
                                                                },
                                                                "city": {
                                                                    "type": "string"
                                                                },
                                                                "state": {
                                                                    "type": "string"
                                                                },
                                                                "like": {
                                                                    "type": "string"
                                                                },
                                                                "dislike": {
                                                                    "type": "string"
                                                                },
                                                                "favorite_toys": {
                                                                    "type": "string"
                                                                },
                                                                "fears": {
                                                                    "type": "string"
                                                                },
                                                                "favorite_places": {
                                                                    "type": "string"
                                                                },
                                                                "spayed": {
                                                                    "type": "string"
                                                                },
                                                                "birthday": {
                                                                    "type": "date"
                                                                },
                                                                "pictures": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "properties": {
                                                                            "id": {
                                                                                "type": "integer"
                                                                            },
                                                                            "picture": {
                                                                                "type": "string"
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "User not pet owner or credentials not match error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "You are not pet owner.|Credentials don't match."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "field": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "Invalid data"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Logout",
                "description": "Logout method",
                "operationId": "logout",
                "responses": {
                    "200": {
                        "description": "Success message.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "success"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unathenticated user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unathenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}